group experiment status code into subfolder
