+++ extension/src/test/suite/experiments/index.test.ts
@@ -39,7 +39,7 @@ import { BaseWebview } from '../../../webview'
 import { MetricsAndParamsModel } from '../../../experiments/metricsAndParams/model'
 import { MessageFromWebviewType } from '../../../webview/contract'
 import { ExperimentsModel } from '../../../experiments/model'
-import { copyOriginalColors } from '../../../experiments/model/colors'
+import { copyOriginalColors } from '../../../experiments/model/status/colors'
 import { InternalCommands } from '../../../commands/internal'
 import { FileSystemData } from '../../../fileSystem/data'
 import { ExperimentsData } from '../../../experiments/data'
